Ad Copy Generator App/requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pyinstaller==6.1.0
langdetect==1.0.9
tenacity==8.2.3
pyyaml==6.0.1
google-ads==22.1.0

# Phase 2: Performance & Async Processing
celery==5.3.4
redis==5.0.1
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Phase 2: Caching & Performance
cachetools==5.3.2
python-memcached==1.62

# Phase 2 & 3: Testing & Quality
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Phase 3: Documentation & API
flask-restx==1.3.0
swagger-ui-bundle==0.0.9
apispec==6.3.0
marshmallow==3.20.1

# Phase 3: Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
